Improve navigation with events and modes
